refactor(browser): keep wait fix focused
