+++ crates/horizon-browser/src/wait.rs
@@ -42,7 +42,6 @@ pub(crate) enum WaitStop {
     /// The browser backend stopped before the wait reached a terminal state.
     BrowserUnavailable,
 }
-
 pub(crate) type WaitResult = Result<BrowserControlValue, BrowserControlFailure>;
 
 /// What one observation concluded.
@@ -337,4 +336,416 @@ fn observation_failure_is_transient(failure: &BrowserControlFailure) -> bool {
 }
 
 #[cfg(test)]
-mod tests;
+mod tests {
+    use super::*;
+    use crate::BrowserControlAction;
+
+    fn node(visible: bool) -> BrowserNode {
+        BrowserNode {
+            reference: String::new(),
+            role: "paragraph".to_string(),
+            name: String::new(),
+            text: "late".to_string(),
+            visible,
+            enabled: true,
+            bounds: None,
+        }
+    }
+
+    fn pending(state: SelectorState, timeout_millis: u64, now: Instant) -> PendingWait {
+        let request = AgentAction {
+            action_id: "wait-1".to_string(),
+            actor: "horizon:agent".to_string(),
+            requested_at_millis: 0,
+            action: BrowserControlAction::WaitForSelector {
+                selector: "#late".to_string(),
+                state,
+                timeout_millis: Some(timeout_millis),
+            },
+        };
+        PendingWait::new(
+            request,
+            "#late".to_string(),
+            state,
+            Some(timeout_millis),
+            Duration::ZERO,
+            7,
+            now,
+        )
+    }
+
+    fn scan() -> serde_json::Value {
+        serde_json::json!({ "nodes": [] })
+    }
+
+    fn done(observation: Observation) -> WaitResult {
+        match observation {
+            Observation::Done(result) => result,
+            other => panic!("expected a completed wait, got {other:?}"),
+        }
+    }
+
+    fn released(wait: &PendingWait, nodes: Vec<BrowserNode>, now: Instant) -> WaitOutcome {
+        match wait.outcome(7, 1, nodes, now) {
+            BrowserControlValue::Wait { wait } => wait,
+            other => panic!("unexpected value {other:?}"),
+        }
+    }
+
+    #[test]
+    fn a_wait_settles_when_the_selector_reaches_the_requested_state() {
+        let now = Instant::now();
+        let mut wait = pending(SelectorState::Visible, 5_000, now);
+        assert!(wait.poll_due(now));
+        assert!(
+            matches!(wait.observe(7, &[], None, now), Observation::Waiting),
+            "no match keeps waiting"
+        );
+        assert!(
+            !wait.poll_due(now + Duration::from_millis(50)),
+            "observations are paced"
+        );
+        assert!(wait.poll_due(now + WAIT_POLL_INTERVAL));
+        assert!(
+            matches!(
+                wait.observe(7, &[node(false)], None, now + Duration::from_millis(100)),
+                Observation::Waiting
+            ),
+            "a hidden match does not satisfy visible"
+        );
+        assert!(matches!(
+            wait.observe(7, &[node(true)], None, now + Duration::from_millis(1_500)),
+            Observation::Satisfied
+        ));
+        // Released later, after the coordination refresh: the elapsed time
+        // still ends at the observation that met the condition.
+        let satisfied = released(&wait, vec![node(true)], now + Duration::from_millis(1_900));
+        assert_eq!(satisfied.state, SelectorState::Visible);
+        assert_eq!(satisfied.polls, 3);
+        assert_eq!(satisfied.elapsed_millis, 1_500);
+        assert_eq!(satisfied.nodes.len(), 1);
+
+        let mut hidden = pending(SelectorState::Hidden, 5_000, now);
+        assert!(matches!(
+            hidden.observe(7, &[node(true)], None, now),
+            Observation::Waiting
+        ));
+        assert!(
+            matches!(hidden.observe(7, &[], None, now), Observation::Satisfied),
+            "an empty match set is hidden"
+        );
+
+        let mut present = pending(SelectorState::Present, 5_000, now);
+        assert!(matches!(
+            present.observe(7, &[node(false)], None, now),
+            Observation::Satisfied
+        ));
+        assert_eq!(released(&present, vec![node(false)], now).polls, 1);
+    }
+
+    #[test]
+    fn timeouts_and_cancellations_are_typed_failures() {
+        let now = Instant::now();
+        let mut wait = pending(SelectorState::Present, 1_000, now);
+        assert!(matches!(wait.observe(7, &[], None, now), Observation::Waiting));
+        assert!(wait.tick(7, now + Duration::from_millis(999)).is_none());
+        let timeout = wait
+            .tick(7, now + Duration::from_secs(1))
+            .expect("settled")
+            .expect_err("timed out");
+        assert_eq!(timeout.code, "wait_timeout");
+        assert!(timeout.message.contains("1 observations"), "{}", timeout.message);
+
+        let invalidated = wait.tick(8, now).expect("settled").expect_err("cancelled");
+        assert_eq!(invalidated.code, "wait_navigation_invalidated");
+        assert_eq!(
+            wait.stopped(WaitStop::OwnershipLost, now).expect_err("cancelled").code,
+            "wait_ownership_lost"
+        );
+        assert_eq!(
+            wait.stopped(WaitStop::HandoffPending, now).expect_err("cancelled").code,
+            "wait_handoff_pending"
+        );
+        assert_eq!(
+            wait.stopped(WaitStop::Superseded, now).expect_err("cancelled").code,
+            "wait_superseded"
+        );
+        assert_eq!(
+            wait.stopped(WaitStop::BrowserUnavailable, now)
+                .expect_err("cancelled")
+                .code,
+            "browser_unavailable"
+        );
+
+        let transient = BrowserControlFailure::new("javascript_error", "Execution context was destroyed");
+        assert!(
+            wait.observe_failure(transient, now).is_none(),
+            "transient failures are retried"
+        );
+        let invalid = BrowserControlFailure::new("invalid_selector", "bad selector");
+        assert_eq!(
+            wait.observe_failure(invalid, now)
+                .expect("settled")
+                .expect_err("failed")
+                .code,
+            "invalid_selector"
+        );
+
+        let queued = PendingWait::new(
+            wait.request.clone(),
+            "#late".to_string(),
+            SelectorState::Present,
+            Some(1_000),
+            Duration::from_millis(400),
+            7,
+            now,
+        );
+        assert!(queued.tick(7, now + Duration::from_millis(599)).is_none());
+        assert!(
+            queued.tick(7, now + Duration::from_millis(600)).is_some(),
+            "queue latency counts against the bound"
+        );
+    }
+
+    #[test]
+    fn the_condition_counts_every_match_and_the_outcome_is_capped() {
+        let now = Instant::now();
+        let mut nodes: Vec<BrowserNode> = (0..25).map(|_| node(false)).collect();
+        nodes.push(node(true));
+        let mut visible = pending(SelectorState::Visible, 1_000, now);
+        assert!(
+            matches!(visible.observe(7, &nodes, None, now), Observation::Satisfied),
+            "a visible match beyond the cap counts"
+        );
+        assert_eq!(
+            released(&visible, nodes.clone(), now).nodes.len(),
+            WAIT_MAX_RESULTS as usize
+        );
+        let mut hidden = pending(SelectorState::Hidden, 1_000, now);
+        assert!(
+            matches!(hidden.observe(7, &nodes, None, now), Observation::Waiting),
+            "a visible match beyond the cap keeps hidden unsatisfied"
+        );
+
+        // With the scan's summary, matches beyond the returned cap decide too:
+        // 300 matches of which only the 251st is visible.
+        let beyond = ScanSummary {
+            matched: 300,
+            visible: 1,
+        };
+        let capped: Vec<BrowserNode> = (0..250).map(|_| node(false)).collect();
+        let mut visible_beyond = pending(SelectorState::Visible, 1_000, now);
+        assert!(
+            matches!(
+                visible_beyond.observe(7, &capped, Some(beyond), now),
+                Observation::Satisfied
+            ),
+            "a visible match beyond the scan cap satisfies visible"
+        );
+        let mut hidden_beyond = pending(SelectorState::Hidden, 1_000, now);
+        assert!(
+            matches!(
+                hidden_beyond.observe(7, &capped, Some(beyond), now),
+                Observation::Waiting
+            ),
+            "a visible match beyond the scan cap keeps hidden unsatisfied"
+        );
+        let mut present_none = pending(SelectorState::Present, 1_000, now);
+        assert!(matches!(
+            present_none.observe(7, &[], Some(ScanSummary { matched: 0, visible: 0 }), now),
+            Observation::Waiting
+        ));
+    }
+
+    #[test]
+    fn an_observation_from_another_page_generation_invalidates_the_wait() {
+        let now = Instant::now();
+        let mut pending = pending(SelectorState::Present, 1_000, now);
+        let result = done(pending.observe(7 + 1, &[node(true)], None, now));
+        assert_eq!(result.expect_err("invalidated").code, "wait_navigation_invalidated");
+    }
+
+    #[test]
+    fn observations_are_budgeted_and_rechecked_against_the_deadline() {
+        let now = Instant::now();
+        let pending_wait = pending(SelectorState::Present, 1_000, now);
+        assert_eq!(pending_wait.observation_budget(now), Duration::from_secs(1));
+        assert_eq!(
+            pending_wait.observation_budget(now + Duration::from_millis(900)),
+            MIN_OBSERVATION_BUDGET,
+            "the last observation still gets a tick"
+        );
+        let long = pending(SelectorState::Present, 60_000, now);
+        assert_eq!(long.observation_budget(now), MAX_OBSERVATION_BUDGET);
+
+        // A query that returned after the bound is a timeout even when it saw
+        // the element, and a transient failure after the bound is one too.
+        let late = now + Duration::from_millis(1_001);
+        let mut overran = pending(SelectorState::Present, 1_000, now);
+        let failure = done(overran.observe(7, &[node(true)], None, late)).expect_err("timed out");
+        assert_eq!(failure.code, "wait_timeout");
+        assert!(
+            failure.message.contains("within the 1000 ms bound (elapsed 1001 ms"),
+            "the message names the configured bound and the elapsed time separately: {}",
+            failure.message
+        );
+        let mut failed_late = pending(SelectorState::Present, 1_000, now);
+        let failure = BrowserControlFailure::new("protocol_error", "evaluate timed out");
+        assert_eq!(
+            failed_late
+                .observe_failure(failure, late)
+                .expect("settled")
+                .expect_err("timed out")
+                .code,
+            "wait_timeout"
+        );
+        let mut failed_early = pending(SelectorState::Present, 1_000, now);
+        assert!(
+            failed_early
+                .observe_failure(
+                    BrowserControlFailure::new("protocol_error", "Execution context was destroyed"),
+                    now
+                )
+                .is_none(),
+            "a transient failure inside the bound is retried"
+        );
+    }
+
+    #[test]
+    fn only_retryable_failures_are_retried_and_only_a_few_times() {
+        let now = Instant::now();
+        let mut pending_wait = pending(SelectorState::Present, 10_000, now);
+        assert!(
+            pending_wait
+                .observe_failure(
+                    BrowserControlFailure::new("javascript_error", "Execution context was destroyed"),
+                    now
+                )
+                .is_none(),
+            "a replaced execution context is retried"
+        );
+        assert!(
+            pending_wait
+                .observe_failure(
+                    BrowserControlFailure::new("protocol_error", "Runtime.evaluate timed out after 250 ms"),
+                    now
+                )
+                .is_none(),
+            "an observation timeout is retried"
+        );
+        assert!(
+            pending_wait
+                .observe_failure(
+                    BrowserControlFailure::new(
+                        "javascript_error",
+                        "WebDriver HTTP I/O: Resource temporarily unavailable (os error 11)"
+                    ),
+                    now
+                )
+                .is_none(),
+            "a Unix read deadline on a bounded WebDriver observation is retried"
+        );
+        for (code, message) in [
+            ("protocol_error", "CDP connection closed"),
+            ("protocol_error", "no page session for Runtime.evaluate"),
+            ("javascript_error", "ReferenceError: foo is not defined"),
+            ("invalid_result", "not a node list"),
+        ] {
+            let mut fresh = pending(SelectorState::Present, 10_000, now);
+            let surfaced = fresh
+                .observe_failure(BrowserControlFailure::new(code, message), now)
+                .expect("settled");
+            assert_eq!(surfaced.expect_err("failed").code, code, "{message} is permanent");
+        }
+
+        let mut persistent = pending(SelectorState::Present, 10_000, now);
+        for _ in 1..MAX_TRANSIENT_FAILURES {
+            assert!(
+                persistent
+                    .observe_failure(BrowserControlFailure::new("protocol_error", "evaluate timed out"), now)
+                    .is_none()
+            );
+        }
+        let surfaced = persistent
+            .observe_failure(BrowserControlFailure::new("protocol_error", "evaluate timed out"), now)
+            .expect("settled");
+        assert_eq!(surfaced.expect_err("failed").code, "protocol_error");
+    }
+
+    #[test]
+    fn a_failed_release_check_is_retried_without_the_wait_deadline() {
+        let now = Instant::now();
+        let mut held = pending(SelectorState::Present, 1_000, now);
+        assert!(matches!(
+            held.observe(7, &[node(true)], None, now),
+            Observation::Satisfied
+        ));
+        held.defer(scan(), 3);
+        let late = now + Duration::from_secs(3);
+        for _ in 1..MAX_TRANSIENT_FAILURES {
+            assert!(
+                held.release_check_failure(BrowserControlFailure::new(
+                    "javascript_error",
+                    "WebDriver HTTP I/O: Resource temporarily unavailable (os error 11)"
+                ))
+                .is_none(),
+                "a slow identity read past the deadline is retried, not timed out"
+            );
+        }
+        assert!(
+            held.tick(7, late).is_none(),
+            "the held match still ignores the deadline"
+        );
+        assert_eq!(held.polls, 1, "release checks are not observations");
+        let surfaced = held
+            .release_check_failure(BrowserControlFailure::new("javascript_error", "evaluate timed out"))
+            .expect("settled");
+        assert_eq!(surfaced.expect_err("failed").code, "javascript_error");
+
+        let mut permanent = pending(SelectorState::Present, 1_000, now);
+        assert!(matches!(
+            permanent.observe(7, &[node(true)], None, now),
+            Observation::Satisfied
+        ));
+        permanent.defer(scan(), 3);
+        let surfaced = permanent
+            .release_check_failure(BrowserControlFailure::new("invalid_result", "no document identity"))
+            .expect("settled");
+        assert_eq!(surfaced.expect_err("failed").code, "invalid_result");
+        assert!(RELEASE_CHECK_BUDGET > MIN_OBSERVATION_BUDGET);
+    }
+
+    #[test]
+    fn a_satisfied_scan_is_released_only_after_a_later_signal_refresh() {
+        let now = Instant::now();
+        // The deferred scan survives the deadline passing meanwhile; a
+        // navigation still invalidates it; and it is released only under a
+        // later coordination-read epoch than it was observed under.
+        let mut deferred = pending(SelectorState::Present, 1_000, now);
+        assert!(matches!(
+            deferred.observe(7, &[node(true)], None, now),
+            Observation::Satisfied
+        ));
+        deferred.defer(scan(), 3);
+        assert!(
+            deferred.tick(7, now + Duration::from_secs(2)).is_none(),
+            "the deadline does not apply to a held result"
+        );
+        assert!(
+            deferred.tick(8, now + Duration::from_secs(2)).is_some(),
+            "a navigation still invalidates it"
+        );
+        assert!(
+            deferred.take_deferred(3).is_none(),
+            "no refresh has run since the observation"
+        );
+        assert!(deferred.take_deferred(4).is_some(), "a later refresh releases it");
+        assert!(deferred.take_deferred(5).is_none());
+
+        // A wait blocked by an in-flight navigation remembers it.
+        let mut blocked = pending(SelectorState::Present, 1_000, now);
+        assert!(!blocked.blocked_by_navigation());
+        blocked.block_for_navigation();
+        assert!(blocked.blocked_by_navigation());
+    }
+}
