+++ docs/architecture/maintainability.md
@@ -33,8 +33,7 @@ back into large multi-purpose modules.
   lives in `navigation.rs` so both drivers settle the same typed outcome.
   Selector waits follow the same shape: the backend-neutral `PendingWait`
   state machine lives in `wait.rs`, and each driver's observation loop glue
-  in `session/wait.rs` and `webdriver/session/wait.rs`; focused wait-state
-  tests live under `wait/tests/` rather than dominating the state-machine file.
+  in `session/wait.rs` and `webdriver/session/wait.rs`.
 - `webdriver/session.rs` orchestrates Firefox and Safari. Host coordination
   belongs in `webdriver/session/coordination.rs`, synchronous navigation
   outcomes in `webdriver/session/navigation.rs`, and HTTP, action translation,
